requirements.txt
matplotlib>=3.8.0
numpy>=1.26.0
pandas>=2.1.0
polars>=0.20.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
plotly>=5.20.0
streamlit>=1.31.0
fastapi>=0.110.0
uvicorn>=0.27.0
joblib>=1.3.0
pyyaml>=6.0.1
sqlalchemy>=2.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
pymysql>=1.1.0
